+++ CHANGELOG.md
@@ -15,6 +15,7 @@ nav_order: 1
 - Add OpenSearch external integration endpoint
 - Add `aiven_pg_user` import example to docs
 - Extend converter for the service user configuration options `ip_filter` object format
+- Fix `aiven_azure_privatelink_connection_approval` import
 
 ## [3.8.0] - 2022-09-30